fix: authOptions 파일에 auth 로직 이동 route는 라우터 기능만 하도록 수정
